Update playback status in playlist view.
